macros: Move additional utility function into util module
